Add Link related events

+++ core/src/main/java/me/mastercapexd/auth/messenger/commands/AccountEnterDeclineCommand.java
@@ -5,9 +5,11 @@ import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 import com.bivashy.auth.api.AuthPlugin;
+import com.bivashy.auth.api.event.AccountLinkEnterDeclineEvent;
 import com.bivashy.auth.api.link.LinkType;
 import com.bivashy.auth.api.link.user.entry.LinkEntryUser;
 
+import io.github.revxrsal.eventbus.EventBus;
 import me.mastercapexd.auth.link.LinkCommandActorWrapper;
 import me.mastercapexd.auth.messenger.commands.annotation.CommandKey;
 import revxrsal.commands.annotation.Default;
@@ -20,9 +22,11 @@ public class AccountEnterDeclineCommand implements OrphanCommand {
     public static final String CONFIGURATION_KEY = "enter-decline";
     @Dependency
     private AuthPlugin plugin;
+    @Dependency
+    private EventBus eventBus;
 
     @DefaultFor("~")
-    public void onDecline(LinkCommandActorWrapper actorWrapper, LinkType linkType, @Default("all") String acceptPlayerName) {
+    public void onDecline(LinkCommandActorWrapper actorWrapper, LinkType linkType, @Default("all") String declinePlayerName) {
         List<LinkEntryUser> accounts = plugin.getLinkEntryBucket().getLinkUsers(entryUser -> {
             if (!entryUser.getLinkType().equals(linkType))
                 return false;
@@ -35,19 +39,22 @@ public class AccountEnterDeclineCommand implements OrphanCommand {
             if (confirmationSecondsPassed.getSeconds() > linkType.getSettings().getEnterSettings().getEnterDelay())
                 return false;
 
-            if (!acceptPlayerName.equals("all")) // If player not default value
-                return entryUser.getAccount().getName().equalsIgnoreCase(acceptPlayerName); // Check if entryUser name
-            // equals to accept player
+            if (!declinePlayerName.equals("all"))
+                return entryUser.getAccount().getName().equalsIgnoreCase(declinePlayerName);
             return true;
         });
         if (accounts.isEmpty()) {
             actorWrapper.reply(linkType.getLinkMessages().getMessage("enter-no-accounts"));
             return;
         }
-        accounts.forEach((entryUser) -> {
+        accounts.forEach((entryUser) -> eventBus.publish(AccountLinkEnterDeclineEvent.class, entryUser.getAccount(), false, linkType, entryUser, entryUser,
+                actorWrapper).thenAccept(result -> {
+            if (result.getEvent().isCancelled())
+                return;
             plugin.getLinkEntryBucket().removeLinkUser(entryUser);
-            entryUser.getAccount().kick(linkType.getServerMessages().getStringMessage("enter-declined", linkType.newMessageContext(entryUser.getAccount())));
+            entryUser.getAccount()
+                    .kick(linkType.getServerMessages().getStringMessage("enter-declined", linkType.newMessageContext(entryUser.getAccount())));
             actorWrapper.reply(linkType.getLinkMessages().getMessage("enter-declined", linkType.newMessageContext(entryUser.getAccount())));
-        });
+        }));
     }
 }
